install.py: Don't try to make rules.txt
Previously, if no_cemu was set, `link_master_mod()` would not make rules.txt, which would send the `while` into an infinite loop.

Also, the entire rules_path section was unnecessary, as `link_master_mod()` makes rules.txt if it's needed, already.

Now the `while` checks for the `content` folder, because that will always be there for every mod.